Make interpolate JS function not require the named parameter.
It is completely unnecessary and annoying to have to specify an additional parameter in Javascript
when using a dictionary to fill in the strings. Switching back and forth between python and js, this
can be often times forgotton, with bad results.

With this fix, compatibility is maintained with all current utilizations.